Move mime code to separate package
